changes file layout for better clarity
