+++ src/data/mechanisims/mechanism.js
@@ -1,0 +1,2 @@
+export const mechanism = {
+}
